Cromper mkst (#1690)
* Fix diff cell background color setting (#1681)

* Remove 'include_presets' from /api/platform endpoint (#1683)

* Remove 'include_presets' from /api/platform endpoint

* 🐶 ruff ruff

* Remove broken test

* Fix 500 when creating scratches with a preset that have libraries (#1685)

* Fix 500 when creating scratches with a preset that have libraries

* Allow users to specify libraries when creating scratches

* Add validation to user-supplied libraries

* 🐕

* mypy, yourpy, ourpy

* woof

* Remove plausible (#1686)

* Remove 'num_scratches' from platform endpoint (#1688)

* Split handlers out of main.py

* Move wrappers into wrappers/ dir

* Rudimentary 'Library' support in cromper

* ruff

* Band-aid fix for /platform

* PR feedback

* Pass config to Handlers instead of using self.application...

* 'Platforms' instance + few bugfixes

* remove /library endpoint from django

* fixup presets with libraries

* clean up dead func

* missed one

* Make cromper urls more explicit

* Remove duplicated Library class
