+++ backend/coreapp/views/scratch.py
@@ -30,7 +30,7 @@ from ..filters.search import NonEmptySearchFilter
 from ..middleware import Request
 from ..models.best_fork import update_best_forks_for_scratch
 from ..models.preset import Preset
-from ..models.scratch import Asm, Assembly, Scratch
+from ..models.scratch import Asm, Assembly, Scratch, Library
 from ..serializers import (
     ClaimableScratchSerializer,
     ScratchCompileSerializer,
@@ -91,13 +91,16 @@ def cache_object(platform_arch: str, file: File[Any]) -> Assembly:
 
 def compile_scratch(scratch: Scratch, context: str | None = None) -> CompilationResult:
     try:
+        libraries = [
+            l.to_json() if isinstance(l, Library) else l for l in scratch.libraries
+        ]
         cromper_client = get_cromper_client()
         result = cromper_client.compile_code(
             compiler_id=scratch.compiler,
             compiler_flags=scratch.compiler_flags,
             code=scratch.source_code,
             context=scratch.context,
-            libraries=scratch.libraries,
+            libraries=libraries,
         )
         return CompilationResult(result["elf_object"], result["errors"])
     except (CompilationError, APIException) as e:
@@ -210,9 +213,7 @@ def create_scratch(
 
     cromper_client = get_cromper_client()
     compiler = cromper_client.get_compiler_by_id(data["compiler"])
-    platform_id = data.get("platform", compiler.platform)
-
-    platform = cromper_client.get_platform_by_id(platform_id)
+    platform = data.get("platform", compiler.platform)
 
     target_asm: str = data.get("target_asm", "")
     target_obj: File[Any] | None = data.get("target_obj")
@@ -225,7 +226,7 @@ def create_scratch(
     else:
         asm = get_db_asm(target_asm)
         cromper_client = get_cromper_client()
-        asm_result = cromper_client.assemble_asm(platform_id, asm)
+        asm_result = cromper_client.assemble_asm(platform.id, asm)
 
         # Create Assembly object from cromper response
         assembly, _ = Assembly.objects.get_or_create(
