+++ backend/docker.dev.env
@@ -6,4 +6,3 @@ MEDIA_URL=http://localhost/media/
 
 # Cromper settings
 CROMPER_URL=http://cromper:8888
-
