+++ cromper/cromper/wrappers/diff_wrapper.py
@@ -11,11 +11,10 @@ from dataclasses import dataclass
 
 import diff as asm_differ
 
-from .platforms import Platform
-from .flags import ASMDIFF_FLAG_PREFIX
-from .error import AssemblyError, DiffError, NmError, ObjdumpError
-from .sandbox import Sandbox
-from .wrapper_result import DiffResult
+from ..platforms import Platform
+from ..flags import ASMDIFF_FLAG_PREFIX
+from ..error import AssemblyError, DiffError, NmError, ObjdumpError
+from ..sandbox import Sandbox
 
 logger = logging.getLogger(__name__)
