+++ cromper/tests/test_compilation.py
@@ -4,7 +4,7 @@ Test compilation functionality in cromper.
 
 import unittest
 
-from cromper.compiler_wrapper import CompilationResult
+from cromper.wrappers.compiler_wrapper import CompilationResult
 from cromper.compilers import (
     GCC281PM,
     IDO53,
@@ -13,7 +13,7 @@ from cromper.compilers import (
     WATCOM_105_C,
 )
 from cromper import platforms
-from cromper.diff_wrapper import DiffWrapper
+from cromper.wrappers.diff_wrapper import DiffWrapper
 from cromper.flags import Language
 
 from .common import CromperTestCase, requiresCompiler, compilers
@@ -223,7 +223,7 @@ class CompilerFilteringTests(CromperTestCase):
 
     def test_filter_compiler_flags(self) -> None:
         """Test that compiler flag filtering works correctly."""
-        from cromper.compiler_wrapper import CompilerWrapper
+        from cromper.cromper.wrappers.compiler_wrapper import CompilerWrapper
 
         test_cases = [
             # (input_flags, expected_output)
