+++ frontend/src/lib/api/request.ts
@@ -8,6 +8,9 @@ if (!API_BASE) {
     throw new Error("No API_BASE set");
 }
 
+const CROMPER_BASE =
+    process.env.INTERNAL_CROMPER_BASE ?? process.env.NEXT_PUBLIC_API_BASE;
+
 type Json = any;
 
 const commonOpts: RequestInit = {
@@ -41,9 +44,16 @@ export class RequestFailedError extends Error {
     }
 }
 
+function isCromperUrl(url: string) {
+    if (url.startsWith("/platform")) return true;
+    if (url.startsWith("/compiler")) return true;
+    if (url.startsWith("/library")) return true;
+    return false;
+}
+
 export function normalizeUrl(url: string) {
     if (url.startsWith("/")) {
-        url = API_BASE + url;
+        url = (isCromperUrl(url) ? CROMPER_BASE : API_BASE) + url;
     }
     return url;
 }
