+++ nginx/development/default.conf
@@ -16,7 +16,7 @@ server {
     location / {
         try_files $uri @proxy_frontend;
     }
-    location ~ ^/api/(compiler|platform) {
+    location ~ ^/api/(compiler|platform|library) {
         # drop /api from url
         rewrite ^/api/(.*)$ /$1 break;
         try_files $uri @proxy_cromper;
