+++ nginx/production/default.conf
@@ -94,7 +94,7 @@ server {
         try_files /dummy.html @proxy_frontend;
     }
 
-    location ~ ^/api/(compiler|platform) {
+    location ~ ^/api/(compiler|platform|library) {
         # drop /api from url
         rewrite ^/api/(.*)$ /$1 break;
         try_files $uri @proxy_cromper;
